+++ CHANGELOG.md
@@ -37,3 +37,7 @@
 * Add `Customer#complete()` and `Attendee#complete()` for determining if all fields are filled out for them
 * Add `TimeSlot#toString(format)` for outputting timeSlot.startsAt + duration to string based on a format argument
 * Improve `Product#constructCalendar` preloading so it continually preloads further months
+
+## Master
+
+* Add `TimeSlot.constructCalendar(timeZone, month)`